Fix typography when text is empty
* Make observing a higher order function instead of using hardcoded closure.
* Add static dictionary where key is the UILabel and value are the attributes for this label. This is a hack to have a stored property in extension without using objc.
* Whenever attribute is fetched via getter, check if it should use the static dictionary value. If so, swap the values in label and remove the values from the dictionary.
* Whenever attribute is added to empty string, store it in the static dictionary for a specific label and fix text alignment.
* Whenever attribute is removed, remove it in the dictionary as well.
* Make closure that is triggered on text change do two things
a) _ = self.attributedText -> This code should not do anything as it's only getter from the UIKit, but internal implementation of the UIKit has side-effects that are actually fixing the typography problems! It does not have to be a setter as in previous implementation.
b) _ = self.attributes -> This code is also a getter, but this logic needs to be triggered each time text changes to swap values if previous static dictionary had values (hence, previous value had to be empty string).
NOTE: Code from line "a" is not necessary, as code from line "b" contains code from line "a" anyway. It's added here for the brevity of the iOS Framework side effects